+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## v2.1.0 (2025-06-17)
+
+### Added
+
+- Add a `twig_toolkit_without` filter ([#36](https://github.com/studiometa/twig-toolkit/pull/36), [b2842a2](https://github.com/studiometa/twig-toolkit/commit/b2842a2))
+
+### Fixed
+
+- Fix Twig 3.21 deprecations ([#35](https://github.com/studiometa/twig-toolkit/pull/35), [6687590](https://github.com/studiometa/twig-toolkit/commit/6687590))
+
 ## v2.0.1 (2025-01-27)
 
 ### Changed